add redirection to index.php :arrow_right: article_delete_database.php

+++ database/Article/article_delete_database.php
@@ -3,7 +3,13 @@ session_start();
 include('../../login_database.php');
 include('../connection_database.php');
 
-$delete_article = $bdd->prepare('DELETE FROM article WHERE id_article = :id_article');
-$delete_article->bindParam(':id_article', $_SESSION['id_article']);
-$delete_article->execute();
+if(isset($_SESSION['id_article'])){
+    $delete_article = $bdd->prepare('DELETE FROM article WHERE id_article = :id_article');
+    $delete_article->bindParam(':id_article', $_SESSION['id_article']);
+    $delete_article->execute();
+    header('Location: ../../index.php');
+}else{
+    header('Location: ../../index.php');
+}
+
 ?>
